Fix ktlint formatting issues
